renamed endpoint

+++ src/main/java/de/vwgis/kafkatitanixexercise/model/BasicPassenger.java
@@ -1,9 +1,11 @@
 package de.vwgis.kafkatitanixexercise.model;
 
+import java.math.BigDecimal;
+
 public interface BasicPassenger {
     Integer getPassengerId();
 
-    Double getFare();
+    BigDecimal getFare();
 
     Boolean getSurvived();
